Changed variables to double and started on gaussian distribution generator

+++ include/anomaly.h
@@ -1,6 +1,6 @@
 #ifndef ANOMALY_H
 #define ANOMALY_H
 
-int is_anomaly(float current, float avg);
+int is_anomaly(double current, double avg);
 
 #endif
